Add ctrl-f and f3 shortcuts for finding within page
I initiated a pull request a few weeks ago, but due to a combination of mistakes with version
control and functionality of my PR, it was abandoned.  This PR should be merged for users of
a hardware keyboard, so that users can easily find within page.  This can also serve as the
basis of adding keyboard functionality for other features, but for now, this pull request
will only add finding within page functionality.  Ctrl-f is fairly ubiquitous, but F3 is used
in Chrome, so certain users (e.g. those with limited mobility or children with small hands) can
have full access to the convenience of keyboard shortcuts.  To test this,